+++ src/uqtestfuns/core/prob_input/univariate_distributions/exponential.py
@@ -1,32 +1,32 @@
 """
-Module with routines involving the normal (Gaussian) probability distribution.
+Module with routines involving the exponential probability distribution.
 
-The normal distribution in UQTestFuns is parametrized by two parameters:
-mu and sigma, the mean and standard deviation, respectively.
+The exponential distribution in UQTestFuns is parametrized by a single
+parameter: the rate parametere.
 
 The underlying implementation is based on the implementation from scipy.stats.
-In the SciPy convention, the mean (mu) corresponds to the ``loc`` parameter,
-while the standard deviation (sigma) corresponds to the ``scale`` parameter.
+In the SciPy convention, the rate parameter corresponds to the reciprocal
+of the ``scale`` parameter.
 """
 import numpy as np
-from scipy.stats import norm
+from scipy.stats import expon
 
 from .utils import verify_param_nums, postprocess_icdf
 from ....global_settings import ARRAY_FLOAT
 
-DISTRIBUTION_NAME = "normal"
+DISTRIBUTION_NAME = "exponential"
 
-NUM_PARAMS = 2
+NUM_PARAMS = 1
 
 
 def verify_parameters(parameters: ARRAY_FLOAT) -> None:
-    """Verify the parameters of a normal distribution.
+    """Verify the parameters of an exponential distribution.
 
     Parameters
     ----------
     parameters : ARRAY_FLOAT
-        The parameters of a normal distribution
-        (i.e., the mean and standard deviation).
+        The parameter of an exponential distribution
+        (i.e., the rate parameter).
 
     Returns
     ------
@@ -42,65 +42,61 @@ def verify_parameters(parameters: ARRAY_FLOAT) -> None:
     # Verify overall shape
     verify_param_nums(parameters.size, NUM_PARAMS, DISTRIBUTION_NAME)
 
-    if parameters[1] <= 0.0:
+    if parameters[0] <= 0.0:
         raise ValueError(
-            f"The corresponding standard deviation {parameters[1]}"
+            f"The corresponding rate parameter {parameters[0]}"
             f"must be larger than 0.0!"
         )
 
 
 def lower(parameters: ARRAY_FLOAT) -> float:
-    """Get the lower bound of a normal distribution.
+    """Get the lower bound of an exponential distribution.
 
     Parameters
     ----------
     parameters : ARRAY_FLOAT
-        The parameters of a normal distribution.
+        The parameter of an exponential distribution.
 
     Returns
     -------
     float
-        The lower bound of the normal distribution.
+        The lower bound of the exponential distribution.
 
     Notes
     -----
-    - Strictly speaking, a normal distribution is unbounded on the left.
-      However, for numerical reason a lower bound is set.
-    - The lower bound of the normal distribution is chosen such that
-      the probability mass between the lower and upper bound is at least
-      1 - 1e-15.
+    - The parameters are not used in determining the lower bound of
+      the distribution; it must, however, appear for interface consistency.
+      The lower bound of a lognormal distribution is finite; it is 0.0.
     """
-    # -8.222082216130435 is the quantile values with probability of 1e-16
-    # for the standard Normal distribution (mu = 0.0, sigma = 1.0)
-    lower_bound = float(-8.222082216130435 * parameters[1] + parameters[0])
+    lower_bound = 0.0
 
     return lower_bound
 
 
 def upper(parameters: ARRAY_FLOAT) -> float:
-    """Get the upper bound of a normal distribution.
+    """Get the upper bound of an exponential distribution.
 
     Parameters
     ----------
     parameters : ARRAY_FLOAT
-        The parameters of a normal distribution
+        The parameters of an exponential distribution
 
     Returns
     -------
     float
-        The upper bound of the normal distribution.
+        The upper bound of the exponential distribution.
 
     Notes
     -----
-    - Strictly speaking, a normal distribution is unbounded on the right.
+    - Strictly speaking, an exponential distribution is unbounded on the right.
       However, for numerical reason an upper bound is set.
-    - The upper bound of the normal distribution is chosen such that
+    - The upper bound of the exponential distribution is chosen such that
       tbe probability mass between the lower and upper bound is at least
       1 - 1e-15.
     """
-    # 8.209536151601387 is the quantile values with probability of 1-1e-16
-    # for the standard Normal distribution (mu = 0.0, sigma = 1.0)
-    upper_bound = float(8.209536151601387 * parameters[1] + parameters[0])
+    # 36.7368005696771 is the quantile values with probability of 1-1e-16
+    # for the exponential distribution with rate parameter value 1.0
+    upper_bound = float(36.7368005696771 / parameters[0])
 
     return upper_bound
 
@@ -111,23 +107,23 @@ def pdf(
     lower_bound: float,
     upper_bound: float,
 ) -> ARRAY_FLOAT:
-    """Get the PDF values of a normal distribution.
+    """Get the PDF values of an exponential distribution.
 
     Parameters
     ----------
     xx : ARRAY_FLOAT
-        Sample values (realizations) of a normal distribution.
+        Sample values (realizations) of an exponential distribution.
     parameters : ARRAY_FLOAT
-        Parameters of the normal distribution.
+        Parameters of the exponential distribution.
     lower_bound : float
-        Lower bound of the normal distribution.
+        Lower bound of the exponential distribution.
     upper_bound : float
-        Upper bound of the normal distribution.
+        Upper bound of the exponential distribution.
 
     Returns
     -------
     np.ndarray
-        PDF values of the normal distribution on the sample values.
+        PDF values of the exponential distribution on the sample values.
 
     Notes
     -----
@@ -135,7 +131,9 @@ def pdf(
     """
     yy = np.zeros(xx.shape)
     idx = np.logical_and(xx >= lower_bound, xx <= upper_bound)
-    yy[idx] = norm.pdf(xx[idx], loc=parameters[0], scale=parameters[1])
+    rate = parameters[0]
+    scale = 1 / rate
+    yy[idx] = expon.pdf(xx[idx], scale=scale)
 
     return yy
 
@@ -146,23 +144,23 @@ def cdf(
     lower_bound: float,
     upper_bound: float,
 ) -> ARRAY_FLOAT:
-    """Get the CDF values of a normal distribution.
+    """Get the CDF values of an exponential distribution.
 
     Parameters
     ----------
     xx : ARRAY_FLOAT
-        Sample values (realizations) of a normal distribution.
+        Sample values (realizations) of an exponential distribution.
     parameters : ARRAY_FLOAT
-        Parameters of the normal distribution.
+        Parameters of the exponential distribution.
     lower_bound : float
-        Lower bound of the normal distribution.
+        Lower bound of the exponential distribution.
     upper_bound : float
-        Upper bound of the normal distribution.
+        Upper bound of the exponential distribution.
 
     Returns
     -------
     ARRAY_FLOAT
-        CDF values of the normal distribution on the sample values.
+        CDF values of the exponential distribution on the sample values.
 
     Notes
     -----
@@ -178,9 +176,9 @@ def cdf(
 
     yy[idx_lower] = 0.0
     yy[idx_upper] = 1.0
-    yy[idx_rest] = norm.cdf(
-        xx[idx_rest], loc=parameters[0], scale=parameters[1]
-    )
+    rate = parameters[0]
+    scale = 1 / rate
+    yy[idx_rest] = expon.cdf(xx[idx_rest], scale=scale)
 
     return yy
 
@@ -191,18 +189,18 @@ def icdf(
     lower_bound: float,
     upper_bound: float,
 ) -> ARRAY_FLOAT:
-    """Get the inverse CDF values of a normal distribution.
+    """Get the inverse CDF values of an exponential distribution.
 
     Parameters
     ----------
     xx : ARRAY_FLOAT
         Sample values (realizations) in the [0, 1] domain.
     parameters : ARRAY_FLOAT
-        Parameters of a normal distribution.
+        Parameters of an exponential distribution.
     lower_bound : float
-        Lower bound of the normal distribution.
+        Lower bound of the exponential distribution.
     upper_bound : float
-        Upper bound of the normal distribution.
+        Upper bound of the exponential distribution.
 
     Returns
     -------
@@ -216,7 +214,9 @@ def icdf(
     """
 
     # Compute the ICDF
-    yy = norm.ppf(xx, loc=parameters[0], scale=parameters[1])
+    rate = parameters[0]
+    scale = 1 / rate
+    yy = expon.ppf(xx, scale=scale)
 
     # Check if values are within the set bounds
     yy = postprocess_icdf(yy, lower_bound, upper_bound)